Relocate system_scratch.h structs to per-system headers (refs #1196)
The structs in app/components/system_scratch.h were not entity components --
they were per-system scratch buffers parked in registry context. Move each
struct next to its owning system, eliminating the leaked file from
app/components/.

- ScoringSystemScratch, MissRecord, HitRecord -> systems/scoring_system.h
- ObstacleDespawnScratch -> systems/obstacle_despawn_system.h
- PopupDisplayScratch -> systems/popup_display_system.h
- ParticleSystemScratch -> systems/particle_system.h
- MeshChildCleanupScratch -> systems/camera_system.h (existing header)
- WasmSmokeLaneMarkerState -> systems/game_state_system.h

runtime_system_scratch.cpp aggregates the new headers for init/reserve so
the registry-context contract (test_entt_dispatcher_contract) is unchanged.
Tests now include the per-system header that owns the struct they inspect.

Build green, all 901 tests pass, zero warnings.
