+++ app/systems/scoring_system.cpp
@@ -1,10 +1,10 @@
 #include "all_systems.h"
+#include "scoring_system.h"
 #include "../components/game_state.h"
 #include "../components/scoring.h"
 #include "../components/obstacle.h"
 #include "../components/gameplay_intents.h"
 #include "../components/particle.h"
-#include "../components/system_scratch.h"
 #include "../components/transform.h"
 #include "../components/rendering.h"
 #include "../components/rhythm.h"
