+++ tests/test_entt_dispatcher_contract.cpp
@@ -7,7 +7,10 @@
 #include <entt/entt.hpp>
 #include "components/game_state.h"
 #include "components/input_events.h"
-#include "components/system_scratch.h"
+#include "systems/scoring_system.h"
+#include "systems/obstacle_despawn_system.h"
+#include "systems/popup_display_system.h"
+#include "systems/particle_system.h"
 #include "test_helpers.h"
 
 struct GoCounter {
