+++ tests/test_perspective.cpp
@@ -5,7 +5,7 @@
 #include "components/player.h"
 #include "components/rendering.h"
 #include "components/shape_mesh.h"
-#include "components/system_scratch.h"
+#include "systems/camera_system.h"
 #include "components/transform.h"
 #include "entities/obstacle_entity.h"
 #include "rendering/camera_resources.h"
