+++ tests/test_scoring_system.cpp
@@ -2,7 +2,7 @@
 #include <cstddef>
 #include <catch2/matchers/catch_matchers_floating_point.hpp>
 #include "test_helpers.h"
-#include "components/system_scratch.h"
+#include "systems/scoring_system.h"
 
 namespace {
